pyproject.toml
[project]
name = "mcp-gateway"
version = "0.2.0"
description = "Lean MCP gateway for web search and document fetching"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-gateway = "src.gateway:run_stdio"
mcp-gateway-sse = "src.gateway:run_sse"

[tool.uv]
dev-dependencies = ["pytest>=8.0.0"]
